Revert Grizzly 2.2.x integration.
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@52985 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ appserver/web/admin/src/main/java/org/glassfish/web/admin/monitor/HttpServiceStatsProvider.java
@@ -349,7 +349,7 @@ public class HttpServiceStatsProvider implements PostConstruct {
 
     @ProbeListener("glassfish:web:http-service:dataSentEvent")
     public void dataSentEvent(
-        @ProbeParam("size") long size) {
+        @ProbeParam("size") int size) {
         countBytesTransmitted.increment(size);
     }
